Implemented SelectEagerCollectionInitializer
