+++ hibernate-core/src/main/java/org/hibernate/sql/results/graph/collection/internal/AbstractCollectionInitializer.java
@@ -6,16 +6,11 @@
  */
 package org.hibernate.sql.results.graph.collection.internal;
 
+import org.hibernate.collection.spi.PersistentCollection;
 import org.hibernate.engine.spi.CollectionKey;
-import org.hibernate.internal.log.LoggingHelper;
-import org.hibernate.internal.util.StringHelper;
-import org.hibernate.sql.results.graph.collection.CollectionLoadingLogger;
-import org.hibernate.sql.results.jdbc.spi.JdbcValuesSourceProcessingOptions;
-import org.hibernate.sql.results.jdbc.spi.JdbcValuesSourceProcessingState;
 import org.hibernate.metamodel.mapping.PluralAttributeMapping;
 import org.hibernate.query.NavigablePath;
 import org.hibernate.sql.results.graph.collection.CollectionInitializer;
-import org.hibernate.sql.results.graph.DomainResultAssembler;
 import org.hibernate.sql.results.graph.FetchParentAccess;
 import org.hibernate.sql.results.jdbc.spi.RowProcessingState;
 
@@ -26,42 +21,44 @@ import org.hibernate.sql.results.jdbc.spi.RowProcessingState;
  */
 public abstract class AbstractCollectionInitializer implements CollectionInitializer {
 	private final NavigablePath collectionPath;
-	private final PluralAttributeMapping collectionAttributeMapping;
+	protected final PluralAttributeMapping collectionAttributeMapping;
 
-	private final FetchParentAccess parentAccess;
+	protected final FetchParentAccess parentAccess;
 
-	private final boolean selected;
-
-	/**
-	 * refers to the collection's container value - which collection-key?
-	 */
-	private final DomainResultAssembler keyContainerAssembler;
-
-	/**
-	 * refers to the rows entry in the collection.  null indicates that the collection is empty
-	 */
-	private final DomainResultAssembler keyCollectionAssembler;
-
-	// per-row state
-	private Object keyContainerValue;
-	private Object keyCollectionValue;
-
-	private CollectionKey collectionKey;
+	protected PersistentCollection collectionInstance;
+	protected CollectionKey collectionKey;
 
 	@SuppressWarnings("WeakerAccess")
 	protected AbstractCollectionInitializer(
 			NavigablePath collectionPath,
 			PluralAttributeMapping collectionAttributeMapping,
-			FetchParentAccess parentAccess,
-			boolean selected,
-			DomainResultAssembler keyContainerAssembler,
-			DomainResultAssembler keyCollectionAssembler) {
+			FetchParentAccess parentAccess) {
 		this.collectionPath = collectionPath;
 		this.collectionAttributeMapping = collectionAttributeMapping;
 		this.parentAccess = parentAccess;
-		this.selected = selected;
-		this.keyContainerAssembler = keyContainerAssembler;
-		this.keyCollectionAssembler = keyCollectionAssembler;
+	}
+
+	@Override
+	public void resolveKey(RowProcessingState rowProcessingState) {
+		if ( collectionKey != null ) {
+			// already resolved
+			return;
+		}
+
+		final Object parentKey = parentAccess.getParentKey();
+		if ( parentKey != null ) {
+			collectionKey = new CollectionKey(
+					collectionAttributeMapping.getCollectionDescriptor(),
+					parentKey
+			);
+
+			parentAccess.registerResolutionListener( owner -> collectionInstance.setOwner( owner ) );
+		}
+	}
+
+	@Override
+	public PersistentCollection getCollectionInstance() {
+		return collectionInstance;
 	}
 
 	@Override
@@ -78,38 +75,8 @@ public abstract class AbstractCollectionInitializer implements CollectionInitial
 		return getCollectionAttributeMapping();
 	}
 
-	/**
-	 * Are the values for performing this initialization present in the current
-	 * {@link JdbcValuesSourceProcessingState}?
-	 * Or should a separate/subsequent select be performed
-	 *
-	 * todo (6.0) : opportunity for performance gain by batching these selects triggered at the end of processing the JdbcValuesSource
-	 */
-	protected boolean isSelected() {
-		return selected;
-	}
-
 	protected FetchParentAccess getParentAccess() {
 		return parentAccess;
-	}
-
-	/**
-	 * The value of the container/owner side of the collection key (FK).  Identifies the
-	 * owner of the collection
-	 */
-	@SuppressWarnings("WeakerAccess")
-	protected Object getKeyContainerValue() {
-		return keyContainerValue;
-	}
-
-	/**
-	 * The value of the collection side of the collection key (FK).  Identifies
-	 * inclusion in the collection.  Can be null to indicate that the current row
-	 * does not contain any collection values
-	 */
-	@SuppressWarnings("WeakerAccess")
-	protected Object getKeyCollectionValue() {
-		return keyCollectionValue;
 	}
 
 	@Override
@@ -119,78 +86,7 @@ public abstract class AbstractCollectionInitializer implements CollectionInitial
 	}
 
 	@Override
-	public void resolveKey(RowProcessingState rowProcessingState) {
-		if ( collectionKey != null ) {
-			// already resolved
-			return;
-		}
-
-		final CollectionKey loadingKey = rowProcessingState.getCollectionKey();
-		if ( loadingKey != null ) {
-			collectionKey = loadingKey;
-			return;
-		}
-
-		final JdbcValuesSourceProcessingOptions processingOptions = rowProcessingState.getJdbcValuesSourceProcessingState()
-				.getProcessingOptions();
-
-		keyContainerValue = keyContainerAssembler.assemble(
-				rowProcessingState,
-				processingOptions
-		);
-
-		if ( keyCollectionAssembler == null || keyContainerAssembler == keyCollectionAssembler ) {
-			keyCollectionValue = keyContainerValue;
-		}
-		else {
-			keyCollectionValue = keyCollectionAssembler.assemble(
-					rowProcessingState,
-					processingOptions
-			);
-		}
-
-		Object keyContainerValue = getKeyContainerValue();
-		if ( keyContainerValue != null ) {
-			this.collectionKey = new CollectionKey(
-					collectionAttributeMapping.getCollectionDescriptor(),
-					keyContainerValue
-			);
-
-			if ( CollectionLoadingLogger.DEBUG_ENABLED ) {
-				CollectionLoadingLogger.INSTANCE.debugf(
-						"(%s) Current row collection key : %s",
-						StringHelper.collapse( this.getClass().getName() ),
-						LoggingHelper.toLoggableString( getNavigablePath(), this.collectionKey.getKey() )
-				);
-			}
-		}
-		else if ( keyCollectionValue != null ) {
-			this.collectionKey = new CollectionKey(
-					collectionAttributeMapping.getCollectionDescriptor(),
-					keyCollectionValue
-			);
-
-			if ( CollectionLoadingLogger.DEBUG_ENABLED ) {
-				CollectionLoadingLogger.INSTANCE.debugf(
-						"(%s) Current row collection key : %s",
-						StringHelper.collapse( this.getClass().getName() ),
-						LoggingHelper.toLoggableString( getNavigablePath(), this.collectionKey.getKey() )
-				);
-			}
-		}
-		else {
-			this.collectionKey = new CollectionKey(
-					collectionAttributeMapping.getCollectionDescriptor(),
-					parentAccess.getParentKey()
-			);
-		}
-	}
-
-	@Override
 	public void finishUpRow(RowProcessingState rowProcessingState) {
-		keyContainerValue = null;
-		keyCollectionValue = null;
-
 		collectionKey = null;
 	}
 }
