+++ hibernate-core/src/main/java/org/hibernate/sql/results/graph/collection/internal/EagerCollectionAssembler.java
@@ -6,41 +6,17 @@
  */
 package org.hibernate.sql.results.graph.collection.internal;
 
-import org.hibernate.collection.internal.PersistentArrayHolder;
-import org.hibernate.collection.spi.PersistentCollection;
 import org.hibernate.metamodel.mapping.PluralAttributeMapping;
 import org.hibernate.sql.results.graph.collection.CollectionInitializer;
-import org.hibernate.sql.results.graph.DomainResultAssembler;
-import org.hibernate.sql.results.jdbc.spi.JdbcValuesSourceProcessingOptions;
-import org.hibernate.sql.results.jdbc.spi.RowProcessingState;
-import org.hibernate.type.descriptor.java.JavaTypeDescriptor;
 
 /**
  * @author Steve Ebersole
  */
-public class EagerCollectionAssembler implements DomainResultAssembler {
-	private final PluralAttributeMapping fetchedMapping;
-	private final CollectionInitializer initializer;
+public class EagerCollectionAssembler extends AbstractCollectionAssembler {
 
 	public EagerCollectionAssembler(
 			PluralAttributeMapping fetchedMapping,
 			CollectionInitializer initializer) {
-		this.fetchedMapping = fetchedMapping;
-		this.initializer = initializer;
+		super( fetchedMapping, initializer );
 	}
-
-	@Override
-	public Object assemble(RowProcessingState rowProcessingState, JdbcValuesSourceProcessingOptions options) {
-		PersistentCollection collectionInstance = initializer.getCollectionInstance();
-		if ( collectionInstance instanceof PersistentArrayHolder ) {
-			return collectionInstance.getValue();
-		}
-		return collectionInstance;
-	}
-
-	@Override
-	public JavaTypeDescriptor getAssembledJavaTypeDescriptor() {
-		return fetchedMapping.getJavaTypeDescriptor();
-	}
-
 }
